Refactor: move `initialState` to VendorProvider.jsx
